requirements.txt
fastapi
uvicorn
pandas
folium
requests
nomadicml
python-multipart
python-dotenv